first commit
This is a Solidity smart contract called TokenLock that allows users to deposit a specified amount of ERC20 tokens into the contract and receive an ERC1155 non-fungible token (NFT) in return. The NFT is associated with the deposit data (sender address and deposit amount) and locked for a specified period of time (1 hour by default) before it can be withdrawn.

The contract is imported from OpenZeppelin libraries and inherits from ERC1155 and ReentrancyGuard. It defines the owner of the contract, the ERC20 token to be used for deposits, and a mapping that associates each NFT with a lock time.

The deposit function allows users to deposit tokens into the contract. It first checks that the deposit amount is greater than 0 and then prompts the user to approve the contract to spend tokens. After approval, a fee of 1% of the deposit amount is transferred to the contract owner and the remaining tokens are transferred to the contract. An NFT is minted and associated with the deposit data, and the lock time is set for 1 hour. Finally, the NFT ID is logged.

The withdraw function allows users to withdraw their deposited tokens after the lock time has expired. It first checks that the user owns the NFT associated with the deposit and that the lock time has expired. It then retrieves the deposit data from the NFT metadata and transfers the deposited tokens to the user before burning the NFT.

The getData function retrieves the metadata associated with an NFT.

The setToken and setOwner functions allow the contract owner to update the ERC20 token and the contract owner address, respectively.

The contract also includes an event that logs the NFT ID when a deposit is made and a modifier that restricts certain functions to the contract owner.